WIP: move cpu impl to rstsr-native-impl
