+++ rstsr-core/src/device_cpu_serial/mod.rs
@@ -7,7 +7,5 @@ pub mod matmul;
 pub mod operators;
 pub mod reduction;
 
-pub use assignment::*;
 pub use device::*;
 pub use operators::*;
-pub use reduction::*;
