+++ rstsr-core/src/device_cpu_serial/operators/mod.rs
@@ -4,6 +4,3 @@ pub mod op_ternary_arithmetic;
 pub mod op_ternary_common;
 pub mod op_tri;
 pub mod op_with_func;
-pub mod reduction;
-
-pub use op_with_func::*;
