+++ rstsr-core/src/feature_rayon/reduction.rs
@@ -1,4 +1,3 @@
-use crate::device_cpu_serial::reduction::*;
 use crate::feature_rayon::*;
 use crate::prelude_dev::*;
 use core::sync::atomic::{AtomicPtr, Ordering};
