+++ rstsr-native-impl/src/cpu_serial/mod.rs
@@ -1,0 +1,5 @@
+pub mod assignment;
+pub mod matmul_naive;
+pub mod op_tri;
+pub mod op_with_func;
+pub mod reduction;
