+++ rstsr-native-impl/src/prelude_dev.rs
@@ -1,0 +1,7 @@
+pub(crate) use rstsr_common::prelude_dev::*;
+
+pub use crate::cpu_serial::assignment::*;
+pub use crate::cpu_serial::matmul_naive::*;
+pub use crate::cpu_serial::op_tri::*;
+pub use crate::cpu_serial::op_with_func::*;
+pub use crate::cpu_serial::reduction::*;
